--- apps/client/src/entities/home/index.ts
@@ -1,2 +1,0 @@
-export * from './model/types';
-export * from './model/schema';
