--- apps/client/src/entities/home/model/types.ts
@@ -1,24 +1,0 @@
-import { ApiResponse } from '@repo/shared/types';
-
-export interface ApiKeyData {
-  apiKey: string;
-  expiresAt: string;
-  expiresInDays: number;
-  scopes: string[];
-  description: string;
-}
-export type ApiKeyResponse = ApiResponse<ApiKeyData>;
-
-export interface AvailableScopeData {
-  scope: string;
-  description: string;
-}
-
-export interface AvailableScopeGroupData {
-  title: string;
-  scopes: AvailableScopeData[];
-}
-
-export type AvailableScopeListResponse = ApiResponse<{
-  list: AvailableScopeGroupData[];
-}>;
