+++ apps/client/src/shared/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './useDebounce';
-export * from './useScopeSelection';
