+++ apps/client/src/widgets/home/index.ts
@@ -1,9 +1,4 @@
 export { default as ApiKeyHeader } from './ui/ApiKeyHeader';
-export { default as ApiKeyForm } from './ui/ApiKeyForm';
 export { default as ApiKeyDisplay } from './ui/ApiKeyDisplay';
-export * from './model/useCreateApiKey';
-export * from './model/useDeleteApiKey';
-export * from './model/useGetApiKey';
-export * from './model/useUpdateApiKey';
-export * from './model/useGetAvailableScope';
-export * from './model/useRotateApiKey';
+export { ApiKeyForm } from '@repo/shared/ui';
+export * from '@repo/shared/hooks';
