+++ packages/shared/src/hooks/index.ts
@@ -1,4 +1,11 @@
 export * from './useCopyToClipboard';
 export * from './useURLFilters';
+export * from './useScopeSelection';
 export * from './useExchangeToken';
 export * from './useRefreshToken';
+export * from './useCreateApiKey';
+export * from './useGetApiKey';
+export * from './useGetAvailableScope';
+export * from './useRotateApiKey';
+export * from './useUpdateApiKey';
+export * from './useDeleteApiKey';
