+++ packages/shared/src/hooks/useCreateApiKey.ts
@@ -2,16 +2,16 @@ import { authQueryKeys, authUrl, post } from '@repo/shared/api';
 import { UseMutationOptions, useMutation } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 
-import { ApiKeyFormType, ApiKeyResponse } from '@/entities/home';
+import { ApiKeyFormType, ApiKeyResponse } from '../types';
 
-export const useRotateApiKey = (
+export const useCreateApiKey = (
   options?: Omit<
     UseMutationOptions<ApiKeyResponse, AxiosError, ApiKeyFormType>,
     'mutationKey' | 'mutationFn'
   >,
 ) =>
   useMutation({
-    mutationKey: authQueryKeys.postRotateApiKey(),
-    mutationFn: (data: ApiKeyFormType) => post<ApiKeyResponse>(authUrl.postRotateApiKey(), data),
+    mutationKey: authQueryKeys.postApiKey(),
+    mutationFn: (data: ApiKeyFormType) => post<ApiKeyResponse>(authUrl.postApiKey(), data),
     ...options,
   });
