+++ packages/shared/src/hooks/useGetApiKey.ts
@@ -2,7 +2,7 @@ import { authQueryKeys, authUrl, get } from '@repo/shared/api';
 import { minutesToMs } from '@repo/shared/utils';
 import { UseQueryOptions, useQuery } from '@tanstack/react-query';
 
-import { ApiKeyResponse } from '@/entities/home';
+import { ApiKeyResponse } from '../types';
 
 export const useGetApiKey = (
   options?: Omit<UseQueryOptions<ApiKeyResponse>, 'queryKey' | 'queryFn'>,
