+++ packages/shared/src/hooks/useRotateApiKey.ts
@@ -2,16 +2,16 @@ import { authQueryKeys, authUrl, post } from '@repo/shared/api';
 import { UseMutationOptions, useMutation } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 
-import { ApiKeyFormType, ApiKeyResponse } from '@/entities/home';
+import { ApiKeyFormType, ApiKeyResponse } from '../types';
 
-export const useCreateApiKey = (
+export const useRotateApiKey = (
   options?: Omit<
     UseMutationOptions<ApiKeyResponse, AxiosError, ApiKeyFormType>,
     'mutationKey' | 'mutationFn'
   >,
 ) =>
   useMutation({
-    mutationKey: authQueryKeys.postApiKey(),
-    mutationFn: (data: ApiKeyFormType) => post<ApiKeyResponse>(authUrl.postApiKey(), data),
+    mutationKey: authQueryKeys.postRotateApiKey(),
+    mutationFn: (data: ApiKeyFormType) => post<ApiKeyResponse>(authUrl.postRotateApiKey(), data),
     ...options,
   });
