+++ packages/shared/src/ui/index.ts
@@ -12,6 +12,7 @@ export * from './Textarea';
 export * from './Skeleton';
 export * from './CheckBox';
 export * from './Tooltip';
+export { default as ApiKeyForm } from './ApiKeyForm';
 export { default as SignInForm } from './SignInForm';
 export { default as Header } from './Header';
 export { default as FormErrorMessage } from './FormErrorMessage';
